backend-update/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
aiofiles==25.1.0
aiohttp==3.13.3
anyio==4.12.1
azure-storage-blob==12.28.0
certifi==2026.2.25
click==8.3.1
colorama==0.4.6
distro==1.9.0
fastapi==0.135.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
matplotlib==3.10.8
numpy==2.4.3
openai==2.29.0
opencv-python==4.13.0.92
pandas==3.0.1
pdfplumber==0.11.8
pillow==12.0.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.2
python-multipart==0.0.20
reportlab==4.4.10
six==1.17.0
sniffio==1.3.1
starlette==0.52.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.42.0
openpyxl==3.1.5

# ADD GUNICORN FOR RENDER
gunicorn==21.2.0
